Rename: type 폴더에 있던 항목들 삭제
