--- src/main/java/com/zerobase/CafeBom/type/OrderStatus.java
@@ -1,10 +1,0 @@
-package com.zerobase.cafebom.type;
-
-public enum OrderStatus {
-
-  CANCELED,
-  WAITING,
-  COOKING,
-  PREPARED,
-  FINISHED
-}
